2.0
Added functionalities:
User and password system, where you have to use them for enter the program, and to know the account to transfer, and you can change it. This modified the database as it changes the pin system for accessing the program.
Movements system, where every operation such as transfer, deposit and withdraw gets saved into a movement in the database, and later you can see all the movements your account have.